requirements.txt
fastapi
uvicorn[standard]
beautifulsoup4
requests
textblob
fpdf
pytesseract
Pillow
PyMuPDF